Fake db services available
Fake DB services are available for the sake of running the application
without the need of having the database actually running. The are
correctly bound (with simple injector) in order to pass the controller
verification on startup).

Pin is no longer required when filling the DTO for CfCheck. PIN can be
actually null, and it has sense.