Partial work on application submission.
Not yet working.

Anomaly class is invariant.
ToBeSupervised property has been stripped from Application class. Classes to be supervised are detected directly from data through queries.
Application Store service is in charge of assuring that only one active application exists with a given fiscal code and pin. So, after storing an application in the database, marks as deleted all the applications with the same fiscal code and pin.
ApplicationSubmissionResult class give feddbak to the user after an application submission.
A service to find all active applications given a fiscal code is available. It is used on submission of a new application, possibly marking them as deleted.
PinBuilder class is available. It builds a pin as a random string of 6 'simple' chars (ambiguous chars are avoided, e.g. 0, O, 1, l, i, L, J, X, Y, etc.).
The ApplicationController is available. Its POST method stores an application according to domain rules.
SubmitApplication service implements al the domain rules to submit a new application. It reads the database, search for applications with the same fiscal code. Possibly generates a new PIN (when the specified PIN does not match any existing active application's PIN). Stores the application. Returns the feedback to the user.