Async validation improved et al.
Material datePicker better localized, but still not working. To be fixed.

Personal data asyncValidation handles debouncing time and all the stuff to prevent useless calls to the server.

Once shown, the pin box does not disappear anymore.

http client errors are gracefully handled and notified to the user with a 'please try later' message.