Jwt authentication completed
- Backend

JwtAuthentication is performed through a POST to AuthController, passing a Json similar to the following:

{
  "username: "foo",
  "password": "bar"
}

The controller returns a Json formed as the following:

{
  "success": true,
  "jwtToken": "du6ryUyruykw.....uyHgsywgudj",
  "expiryTime": "2018-06-30T10:10:10:000Z"
}

The backend expects a header for all the subsequent requests:

Authorization: Bearer du6ryUyruykw.....uyHgsywgudj

In case of missing header, a "401 Unauthorized" error is returned.

- Frontend

Jwt authentication is managed by an AuthService class. The service calls the backend method and, in case of success, saves the token in the local browser storage area. An interceptor, implemented by te AuthInterceptor class, assures that the token is always sent back to the client in case of every http request. AuthService exposes also service methods like isLoggedIn(), getExpiretionDate(), etc. to feed the authentication process.

Routing is used to redirect user to authentication in case of need.

Fake authentication service is available to allow bypassing the authenticazion during the development of the GUI.

A bootstrap-based navbar has been implemented mainly to allow user logout.

A bootstrap-based login page has been implemented.