Application export in CSV format
Applications can be exported in CSV format.

Backend exposes a method
to get applications as a text/csv content. The method allows to filter
by the full-text search key.
To get the file, a downloadAuthKey must be provided. A second method allows to
get the downloadAuthKey. This last method is protected by JWT
authentication.

The backoffice-frontend implements the export button click method. The logic should
by encapsulated in a dedicated service (to be done).